Project_Build/Backend/requirements.txt
rutokenizer~=0.0.24
deeppavlov~=0.14.0
natasha~=1.4.0
docx2txt~=0.8
Pillow~=8.4.0
pandas~=0.25.3
pdf2image~=1.16.0
torchvision~=0.11.3
torch~=1.10.2
torchsummary~=1.5.1
pytesseract~=0.3.8